chore: move EventArgs classes to separate files
