+++ src/Uno.UI.RemoteControl/ReceivedFrameEventArgs.cs
@@ -1,0 +1,16 @@
+﻿#nullable enable
+
+using System;
+
+namespace Uno.UI.RemoteControl
+{
+	public class ReceivedFrameEventArgs : EventArgs
+	{
+		public ReceivedFrameEventArgs(HotReload.Messages.Frame frame)
+		{
+			Frame = frame;
+		}
+
+		public HotReload.Messages.Frame Frame { get; set; }
+	}
+}
